Enable streaming on device fd (#144)
This ensures the opener of the fd in the kernel module is updated to be a WRITER.
Failing to do so causes the video output capability to be lost when closing the fd.

Cf. https://github.com/floe/backscrub/issues/133
Cf. https://github.com/umlaeute/v4l2loopback/issues/470

Also this commit introduces a proper loopback_free function to release any resources and provide a central point for cleanup before closing the fd.